requirements_inference.txt
opencv-python>=4.5.0
onnxruntime>=1.12.0
numpy>=1.21.0
matplotlib>=3.5.0
Pillow>=8.3.0
